"Configure public path in vite.config.ts for production environment"
The given diff introduces a new configuration in the vite.config.ts file to set the public path based on the environment. If the environment is production, the public path is set to '/REPO_NAME/' otherwise it will be set to '/'. This is done to ensure the correct asset paths in the production build where assets are expected to be served from '/REPO_NAME/'.