Enable backgroundClipText feature and update UI elements, rules
Enabled backgroundClipText feature in Tailwind CSS for future text effects. Updated LandingView to give text elements a 90% opacity for an overlay effect. Expanded rules in 2.1.md to give a more detailed explanation on the prohibition of cheating in games. Redesigned the header in App.vue to have a gradient background and a slight blur for a modern and fresh look. These changes aim to enhance the website's visual appeal and user experience.