Updated UI and introduced new features in views
A new item "Launcher Manuals" was added to the sidebar in the DocumentationView to provide guidance for using the launcher. Modified active key in DocumentationView sidebar for improved UX. Updated PostsView and LandingView for better layout and visibility. In LandingView, also linked buttons to relevant views for better navigation. Added media queries in tailwind.config.js for responsive design. Updated NMenu styling for better UI visibility in App.vue.